+++ pyvergeos/resources/auth_sources.py
@@ -54,7 +54,7 @@ from typing import TYPE_CHECKING, Any
 
 from pyvergeos.exceptions import NotFoundError
 from pyvergeos.filters import build_filter, quote_value
-from pyvergeos.resources.base import ResourceManager, ResourceObject, normalize_fields, split_fields
+from pyvergeos.resources.base import ResourceManager, ResourceObject, split_fields
 
 if TYPE_CHECKING:
     from pyvergeos.client import VergeClient
@@ -183,7 +183,7 @@ class AuthSourceStateManager(ResourceManager["AuthSourceState"]):
 
         # Use default fields if not specified
         if fields:
-            params["fields"] = normalize_fields(fields)
+            params["fields"] = self._projection(fields)
         else:
             params["fields"] = ",".join(self._default_fields)
 
@@ -232,7 +232,7 @@ class AuthSourceStateManager(ResourceManager["AuthSourceState"]):
             "filter": f"state eq {quote_value(key)}",
         }
         if fields:
-            params["fields"] = normalize_fields(fields)
+            params["fields"] = self._projection(fields)
         else:
             params["fields"] = ",".join(self._default_fields)
 
@@ -520,7 +520,7 @@ class AuthSourceManager(ResourceManager["AuthSource"]):
 
         # Use default fields if not specified
         if fields:
-            params["fields"] = normalize_fields(fields)
+            params["fields"] = self._projection(fields)
         else:
             params["fields"] = ",".join(self._default_fields)